# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Complaint Management Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  total_tickets: number;
  open_tickets: number;
  in_progress_tickets: number;
  resolved_today: number;
  high_priority_open: number;
  closed_tickets: number;
  messages_processed_today: number;
  avg_resolution_hours: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface PriorityCount {
  priority: string;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DailyCount {
  /** ISO date string YYYY-MM-DD */
  date: string;
  count: number;
}

export interface WhatsappStatus {
  connected: boolean;
  instance: string;
  state: string;
  api_url: string;
}

export interface ActivityItem {
  id: number;
  ticket_id: number;
  /** created | status_changed | supervisor_action */
  event: string;
  description: string;
  category: string;
  priority: string;
  status: string;
  timestamp: string;
}

export interface Ticket {
  id: number;
  message_text: string;
  category: string;
  priority: string;
  status: string;
  location?: string | null;
  reporter_name?: string | null;
  group_name?: string | null;
  created_at: string;
  updated_at?: string | null;
}

export type TicketStatusUpdateStatus =
  (typeof TicketStatusUpdateStatus)[keyof typeof TicketStatusUpdateStatus];

export const TicketStatusUpdateStatus = {
  open: "open",
  in_progress: "in_progress",
  resolved: "resolved",
  delayed: "delayed",
  closed: "closed",
} as const;

export interface TicketStatusUpdate {
  status: TicketStatusUpdateStatus;
}

export interface TrackedGroup {
  id: number;
  name: string;
  /** WhatsApp JID (e.g. 120363012345678901@g.us) */
  group_id: string;
  description?: string | null;
  enabled: boolean;
  message_count: number;
  created_at: string;
}

export interface TrackedGroupInput {
  name: string;
  group_id: string;
  description?: string | null;
}

export interface TrackedContact {
  id: number;
  name: string;
  /** Phone in E.164 format (e.g. +971501234567) */
  phone: string;
  description?: string | null;
  enabled: boolean;
  message_count: number;
  created_at: string;
}

export interface RawMessage {
  id: number;
  text: string;
  sender: string;
  group_name?: string | null;
  category?: string | null;
  timestamp: string;
}

export interface TrackedContactInput {
  name: string;
  phone: string;
  description?: string | null;
}

export interface TrackedItemPatch {
  enabled?: boolean;
  name?: string;
  description?: string | null;
}

export interface DeletedResponse {
  deleted: boolean;
  id: number;
}

export interface AiChatRequest {
  /** Existing conversation to append to */
  conversation_id: number;
  /** User prompt text */
  message: string;
  /** Optional system prompt override */
  system_prompt?: string | null;
}

export interface AiConversation {
  id: number;
  title: string;
  created_at: string;
  updated_at?: string | null;
}

export interface AiConversationInput {
  title: string;
}

export type AiMessageRole = (typeof AiMessageRole)[keyof typeof AiMessageRole];

export const AiMessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export interface AiMessage {
  id: number;
  conversation_id: number;
  role: AiMessageRole;
  content: string;
  created_at: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type WhatsappQrResponseStatus =
  (typeof WhatsappQrResponseStatus)[keyof typeof WhatsappQrResponseStatus];

export const WhatsappQrResponseStatus = {
  disconnected: "disconnected",
  waiting_scan: "waiting_scan",
  connected: "connected",
} as const;

export interface WhatsappQrResponse {
  /** Base64 PNG data URL of the QR code */
  qr_data_url: string;
  /** Seconds until the QR code expires */
  expires_in: number;
  status: WhatsappQrResponseStatus;
  pairing_token: string;
}

export type WhatsappConnectStatusState =
  (typeof WhatsappConnectStatusState)[keyof typeof WhatsappConnectStatusState];

export const WhatsappConnectStatusState = {
  disconnected: "disconnected",
  waiting_scan: "waiting_scan",
  connected: "connected",
} as const;

export interface WhatsappConnectStatus {
  connected: boolean;
  state: WhatsappConnectStatusState;
  instance: string;
  phone?: string | null;
  connected_at?: string | null;
  api_url: string;
}

export interface TelegramStatus {
  connected: boolean;
  bot_username?: string | null;
  bot_name?: string | null;
  webhook_url?: string | null;
  configured_at?: string | null;
}

export interface TelegramSetupInput {
  /** Telegram bot token from @BotFather (format: 123456:ABC-DEF...) */
  bot_token: string;
  /** Optional HTTPS webhook URL for receiving updates */
  webhook_url?: string | null;
}

export interface TelegramQrResponse {
  qr_data_url: string;
  bot_link: string;
  bot_username: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListTicketsParams = {
  status?: string;
  priority?: string;
  category?: string;
  limit?: number;
};