# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Complaint Management Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List all tickets with optional filters
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: priority
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get a single ticket by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tickets/{id}/status:
    patch:
      operationId: updateTicketStatus
      tags: [tickets]
      summary: Update the status of a ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketStatusUpdate"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /connect/telegram:
    post:
      operationId: setupTelegram
      tags: [connect]
      summary: Configure Telegram bot token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramSetupInput"
      responses:
        "200":
          description: Telegram bot configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramStatus"
        "400":
          description: Invalid bot token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /policies/groups:
    get:
      operationId: listTrackedGroups
      tags: [policies]
      summary: List all tracked WhatsApp groups
      responses:
        "200":
          description: Tracked groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackedGroup"
    post:
      operationId: addTrackedGroup
      tags: [policies]
      summary: Add a WhatsApp group to track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackedGroupInput"
      responses:
        "201":
          description: Group added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedGroup"
        "409":
          description: Group already tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /policies/groups/{id}:
    patch:
      operationId: updateTrackedGroup
      tags: [policies]
      summary: Enable or disable a tracked group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackedItemPatch"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedGroup"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrackedGroup
      tags: [policies]
      summary: Remove a WhatsApp group from tracking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /policies/contacts:
    get:
      operationId: listTrackedContacts
      tags: [policies]
      summary: List all tracked individual contacts
      responses:
        "200":
          description: Tracked contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackedContact"
    post:
      operationId: addTrackedContact
      tags: [policies]
      summary: Add an individual contact to track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackedContactInput"
      responses:
        "201":
          description: Contact added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedContact"
        "409":
          description: Contact already tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /policies/contacts/{id}:
    patch:
      operationId: updateTrackedContact
      tags: [policies]
      summary: Enable or disable a tracked contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackedItemPatch"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackedContact"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTrackedContact
      tags: [policies]
      summary: Remove a contact from tracking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
        message_text:
          type: string
        category:
          type: string
        priority:
          type: string
        status:
          type: string
        location:
          type: string
          nullable: true
        reporter_name:
          type: string
          nullable: true
        group_name:
          type: string
          nullable: true
        created_at:
          type: string
        updated_at:
          type: string
          nullable: true
      required:
        - id
        - message_text
        - category
        - priority
        - status
        - created_at
    TicketStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, resolved, delayed, closed]
      required: [status]
    TrackedGroup:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        group_id:
          type: string
          description: WhatsApp JID (e.g. 120363012345678901@g.us)
        description:
          type: string
          nullable: true
        enabled:
          type: boolean
        message_count:
          type: integer
        created_at:
          type: string
      required: [id, name, group_id, enabled, message_count, created_at]
    TrackedGroupInput:
      type: object
      properties:
        name:
          type: string
        group_id:
          type: string
        description:
          type: string
          nullable: true
      required: [name, group_id]
    TrackedContact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          description: Phone in E.164 format (e.g. +971501234567)
        description:
          type: string
          nullable: true
        enabled:
          type: boolean
        message_count:
          type: integer
        created_at:
          type: string
      required: [id, name, phone, enabled, message_count, created_at]
...
    TrackedContactInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        description:
          type: string
          nullable: true
      required: [name, phone]
    TrackedItemPatch:
      type: object
      properties:
        enabled:
          type: boolean
        name:
          type: string
        description:
          type: string
          nullable: true
    DeletedResponse:
      type: object
      properties:
        deleted:
          type: boolean
        id:
          type: integer
      required: [deleted, id]
...
    TelegramStatus:
      type: object
      properties:
        connected:
          type: boolean
        bot_username:
          type: string
          nullable: true
        bot_name:
          type: string
          nullable: true
        webhook_url:
          type: string
          nullable: true
        configured_at:
          type: string
          nullable: true
      required: [connected]
    TelegramSetupInput:
      type: object
      properties:
        bot_token:
          type: string
          description: "Telegram bot token from @BotFather (format: 123456:ABC-DEF...)"
        webhook_url:
          type: string
          nullable: true
          description: Optional HTTPS webhook URL for receiving updates
      required: [bot_token]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]